requirements.txt
requests
pandas
GitPython